pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photo-backup-organizer"
version = "0.1.0"
description = "Organizes phone backup photos and videos into a curated NAS library"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "Pillow>=10.3.0",
    "exifread>=3.0.0",
    "PyYAML>=6.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "anyio>=4.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
